+++ lib/ice/output/ice_perror.c
@@ -6,9 +6,10 @@
 */
 
 #include <unistd.h>
+
 #include "ice/string.h"
 
-int ice_perror(const char *str)
+ssize_t ice_perror(const char *str)
 {
     return write(2, str, ice_strlen(str));
 }
